clean-up and add missing func
